Move data_m/v/p into common area
Going to use this in the GSAS tab as well.

RE #34587
